+++ src/main/java/com/chone/server/domains/review/infrastructure/jpa/ReviewDetailRepositoryImpl.java
@@ -1,10 +1,11 @@
-package com.chone.server.domains.review.repository;
+package com.chone.server.domains.review.infrastructure.jpa;
 
 import com.chone.server.commons.exception.ApiBusinessException;
 import com.chone.server.domains.review.domain.QReview;
 import com.chone.server.domains.review.domain.Review;
 import com.chone.server.domains.review.dto.response.ReviewDetailResponseDto;
 import com.chone.server.domains.review.exception.ReviewExceptionCode;
+import com.chone.server.domains.review.repository.ReviewDetailSearchRepository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
